Move all utils modules to dedicated folder
