+++ rana_qgis_plugin/utlis/generic.py
@@ -6,7 +6,6 @@ from urllib.parse import parse_qs, urlparse
 from uuid import uuid4
 
 from osgeo import gdal
-from qgis.core import QgsProject, QgsRasterLayer, QgsVectorLayer
 from qgis.PyQt.QtCore import QBuffer, QByteArray, QIODevice, QSettings, Qt
 from qgis.PyQt.QtGui import QFont, QFontMetrics, QImage, QStandardItem
 from slugify import slugify
@@ -20,8 +19,8 @@ from rana_qgis_plugin.auth_3di import get_3di_auth
 from rana_qgis_plugin.simulation.threedi_calls import (
     get_api_client_with_personal_api_token,
 )
-from rana_qgis_plugin.utils_api import get_frontend_settings, get_tenant_details
-from rana_qgis_plugin.utils_settings import hcc_working_dir, rana_cache_dir
+from rana_qgis_plugin.utlis.api import get_frontend_settings, get_tenant_details
+from rana_qgis_plugin.utlis.settings import rana_cache_dir
 
 
 def is_writable(working_dir: str) -> bool:
