+++ rana_qgis_plugin/widgets/utils_file_action.py
@@ -4,11 +4,7 @@ from typing import List, Optional
 from qgis.PyQt.QtCore import QObject, pyqtSignal
 
 from rana_qgis_plugin.constant import SUPPORTED_DATA_TYPES
-from rana_qgis_plugin.utils_api import get_tenant_file_descriptor
-from rana_qgis_plugin.utils_scenario import (
-    get_is_3di_simulation,
-    get_ready_state_from_descriptor,
-)
+from rana_qgis_plugin.utlis.api import get_tenant_file_descriptor
 
 
 class FileAction(Enum):
